+++ client/src/App.js
@@ -4,7 +4,8 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 import SignUp from './pages/SignUp';
 import Login from './pages/Login';
 import Profile from './pages/Profile';
-import Park from './pages/Park';
+import Landing from './pages/Landing';
+import Park from './pages/Park2';
 
 import Navbar from './components/Navbar';
 import Wrapper from './components/Wrapper';
@@ -18,7 +19,7 @@ function App() {
         <Wrapper>
           <Route exact path="/signup" component={SignUp} />
           <Route exact path="/login" component={Login} />
-          <Route exact path="/" component={Login} />
+          <Route exact path="/" component={Landing} />
           <Route exact path="/profile" component={Profile} />
           <Route exact path="/park" component={Park} />
         </Wrapper>
